requirements.txt
asgiref==3.10.0
Django==4.2.25
django-environ==0.12.0
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.15.0
djangorestframework==3.15.2